Make Roots.set_rising and Roots.set_falling special cases of Roots.set.
set_rising and set_falling used to take a boolean flag.  If the flag was true
they would set the array element to Rising and Falling, respectively, but if
the flag was true they set the element to the rather arbitrary value NoRoot.
This change eliminates the flag and makes set_rising and set_falling to partial
applications of set to Rising and Falling.