In quickcheck, simplify the behavior of failing residual functions.
Previously, failing residual functions failed only when a specified time was
reached; however, this led to unpredictable behavior because it's
unpredictable which time values IDA uses to call the residual function.

The semantics is now simplified so that residual functions always fail.  The
brief period in which we experimented with conditional failures suggests we're
not missing out on important bugs.