Update models with model_cmd everywhere and remove code duplication.
shrink_cmd, fixup_cmd, and gen_cmd have to update the model that they are
given.  They contained bits and pieces of model_cmd in them, which was hard to
keep in sync with model_cmd proper.  This has led to a number of bugs.

Duplicates of model_cmd are eliminated and updating models are delegated
entirely to model_cmd.  This introduces a lot of copying, but a quick
performance measurement suggests that the overhead is negligible compared to
the cost of running the OCaml compiler on every test case.