In quickcheck_ida.ml: implement SetStopTime; refactor model_cmd.
model_cmd is split into model_cmd_internal and a wrapper named model_cmd.  In
the former, an exception is not returned as Exn foo, but instead it just raises
foo, which is caught and converted to Exn foo in the wrapper.

The control flow in model_cmd_internal is simpler this way.  Previously, the
handling of Exn in model_cmd made the nesting of conditionals snowball like
this:

  if bad
  then Exn Bar
  else (do_something ();  (* do some of the side effects *)
        if bad'
        then Exn Baz
        else (do_something' ();  (* do the remaining effects *)
              Unit               (* return normal result *)
              ))

In the new interface we just write

  if bad then raise Bar;
  do_something ();
  if bad' raise Baz;
  do_something' ();
  Unit