Bug fix: update last_tret correctly.  Work around bug in set_stop_time.
The model's last_query_time and last_tret were being updated with the same
value during SolveNormal.  This surfaced as a bug with the addition of
set_stop_time.  There was a check in SetStopTime that the stop time is not
before last_tret (i.e. that we're not setting a stop time in the past), but
this was effectively using the value of last_query_time.  This is fixed by
updating last_tret and last_query_time with the correct values.

However, that fix above turned out to be useless, due to a bug -- or perhaps
just bad interface -- in sundials.  Sundials does the same check on stop times,
but it compares the time against its internal time, which can be greater than
the last query time.  This internal time is unpredictable, so we can't run
set_stop_time with predictable results if there has been a call to IDASolve
before.  The generator is modified so that we don't run set_stop_time in that
case.