Auto-generate more stuff; overhaul Makefile; speed up test execution.
quickcheck_ida.ml included wholesale copies of type definitions from ida.mli so
as to derive pretty printers and reifiers (= expr_of's).  This definition had
to be manually kept in sync with ida.mli.  A new utility extract_types is now
responsible for extracting the types automatically.  This is infeasible for
sundials.mli whose types require some special-casing, but at least for ida.mli
this seems to work OK.

The Makefile now generates some new modules using extract_types:

  pprint_ida.ml
  expr_of_ida.ml
  expr_of_ida_model.ml

The Makefile is overhauled to incorporate these new modules, and it is also
changed so that updating .depends works more robustly: generating that file
could fail when the pa_deriving.cmo syntax extension was not compiled, but the
rule is made smarter so that it generates dependencies as much as it can.  When
pa_deriving.cmo is compiled, the .depends is re-generated automatically to get
a complete dependency listing.

With extract_types, the derivation of pretty-printers and reifiers are now
split into different modules.  This allows test cases to be compiled without
linking to any of the expr_of code, which means they don't need to link in
camlp4.  This speeds up testing by roughly a factor of 2x.