Buffer error messages from sundials.
Test code now sets an error handler function which,
 - if --buffer-errors is given, it saves error messages without printing them
 - otherwise prints the messages to standard error as they come in

This serves two purposes.  One is to exercise the error handler mechanism to
make sure it's working.  The other is that, without buffering, error messages
from test cases are displayed not in the [Program Output] section but before
all other outputs (because most of the info is gathered and printed at the
end), which makes the error message easier to miss.  With buffering, the
messages can be placed in the [Program Output] section to get the user's
attention.  If the program crashes, however, the buffered message may never be
displayed, so the buffering behavior is turned on with --buffer-errors only
when we're re-running a test that is known to run to completion.  This happens
when we shrink a test case that doesn't crash but produces incorrect output.
We re-run the last failed test after shrinking, because the last test *tried*
during shrinking is generally not the last one that *failed*.