Abort tests immediately if compilation fails.
In quickcheck, tests used to continue when compilation of generated test
cases failed.  The error messages reported by the compiler and
redirected to quickcheck's stderr would point to test.ml, whereas the
offending code would be found in failed.ml.

Aborting tests immediately instead ensures that the offending code is in
test.ml, as the compilation error says, but at the cost of having the
offending code in different locations depending on the type of error:
test.ml if it's a compilation error, and failed.ml if it's a test
failure.