Always update last_query_time.
In quickcheck_ida, model.last_query_time is used to track the minimum
time at which we can safely query the solution, without running into
unpredictable behavior.  This was not updated when the residual function
or root function failed, which caused an assertion failure.  Update this
state even in the face of an exception.  This state is not used for
anything besides command generation/shrinking anyway.