Fix quickcheck shrinker for ReInit
Don't change the linear solver while fixing up a ReInit command.  The
solver was changed previously because the interface for changing the
solver involved calling lots of functions for setting callbacks and
parameters, and it seemed like a hassle to get all of that right.  With
the updated interface where everything is specified in one go with a
single ADT, this is no longer necessary.  In fact the gratuitous change
in linear solver made it more likely that a shrink would change things
too much so that a bug is no longer triggered, which got us to end up
with a bigger test case than is necessary.