Bug fix: link to Weak.get properly in edge cases
Weak.get is accessed directly from C by one of two techniques:

  - refer directly to the function caml_weak_get() from OCaml stdlib
  - have sundials.ml pass Weak.get to a C function which saves a
    pointer to that function in a generational global root.

The former is preferred, because it's more efficient and goes through
fewer hoops.  This works for 3.12 and 4.01, and judging from OCaml's VCS
history, it should work for any version of OCaml that supplies weak hash
tables to the user.

The latter technique is a fallback to protect against (highly unlikely)
incompatible future changes.  In the implementation of this technique,
callback registration was accidentally omitted, leading to a segfault.
This bug is fixed in this commit.