Bug fix: Fix minor memory leak and corruption.
The generic nvector interface allocated custom blocks that were bigger
than they needed to be, wasting a few words.  This commit fixes the size
calculation.

When malloc fails during creation of a generic nvector, the NVEC_VAL of
the partially constructed nvector was free()'d, but this is incorrect.
The NVEC_VAL points to a memory block managed by OCaml's GC.  This
commit removes the free()s.