Don't use caml_array_length().
This function doesn't exist in OCaml 3.12 and older.  It's also slightly
slower than just using Wosize_val().

The only gotcha is that Wosize_val() doesn't work for doubles of
floats.  Each use of Wosize_val() in lieu of caml_array_length() should
therefore be commented with the expected array type.