Add box plot; allow plotting in subdir
The plotter still gives a bar chart by default, since the box plot looks
a lot more cluttered and is not meaningful with the default
PERF_REPS (you need something like 10 data points to see the variance).
This command gives a box plot:

  make STYLE=boxplot perf.opt.plot

Plotting examples in a specific subdirectory is now as simple as doing

  make perf.opt.plot

in that subdirectory.