Compile C examples locally.
If for any reason the user specified an EXAMPLES directory that contains
only the sources and not the binaries, GNU make would consider the
pattern rule

%.sundials.out: $(EXAMPLESROOT)/$(SUBDIR)/%
                $< > $@

as inapplicable and move on to the built-in rule

%.out: %
       @rm -f @
       cp $< $@

This ends up giving off very cryptic errors to the user.

The new setup will always compile a local version of the examples, so it
works as long as the EXAMPLES directory contains the examples' sources.