Banish Callback.register_exception
Exceptions are instead registered by a nameless scheme, where the
initialization code of OCaml modules call C functions with the desired
exceptions packed in an argument.  This is expected to have a
microscopic improvement on performance, as Callback.register_exception
registers each object in a non-generational global root.  The new scheme
uses generational global roots, so GC has to scan a few dozen roots
less.

As a side effect, all the C-side initialization code in sundials_ml.c is
now collected in one function.