Support side-by-side comparison plot
To get the comparison plot:

- Run make perf.opt.log and copy perf.opt.log somewhere.  Repeat.
- Feed all the *.log files directly to utils/plot.sh.

The side-by-side plots barely fit on the screen, and tends to hit on
bugs in gnuplot's wxt terminal.  This is especially true of the box
plot.  It's best to dump it to a png file with a very large size:

 OUTPUT=foo.png TERMINAL=png SIZE=2400,1600 utils/plot.sh <files...>

and see it in an image viewer.

Single-file plots work as before, apart from minor stylistic changes.