Don't update .reps after reconfigure
configure used to trigger recalculation of .reps files, but this turned
out to be more annoying than expected.  Reconfiguring with
--enable-debug, and then configuring again with optimizations before
running benchmarks is a common workflow, but this has led to
accidentally benchmarking with different sets of .reps.

A quick measurement indicates that CFLAGS has a limited impact on the
number of reps, so this commit removes dependence of .reps on config.

.reps are updated only after distclean, manual removal, or modification
to perf.ml.