Modification to treatement of ids/vartypes.
The IDASetId function specifies whether variables are algebraic or
differential. It must be called before IDASetSuppressArg and
IDACalcIc(IDA_YA_YDP_INIT). As the name ('id') is not well chosen, the ML
interface provided an alias for set_id called set_var_types, but this
complicates the interface with a slight ambiguity and much explanatory
documentation.

This change treats and resolves these issues. Rather than 'var types' we
compromise with 'var id' to clearly link with the underlying library while
still trying to improve the useability of the OCaml interface. We also add a
dynamic check for a prior call to IDASetAlg.

The documentation has not yet been updated.