Sack n_vdestroy; fix nvector finalization
One should never explicitly dispose of objects from the OCaml side.  If
a finalizer is needed, the user can attach one with Gc.finalise to the
payload.

Remove all accesses to the OCaml heap from C-level finalizers.

Streamline function names to be consistent with the explanation in
nvector_ml.h.