Pack callback arguments into records
Leave out BBD, since they're impossible to do cleanly and consistently.
The constraint that {ida,cvode}_impl.ml can't mention parallel types
means the types must be re-declared in bbd modules, but this means we
need to set up a proxy function that unnecessarily copies those records.
It's not just inefficient, but also confusing and complicates the code.

Their requirements are also inconsistent.  Functions in:

 - kinsol_bbd and cvode_bbd take too few arguments to justify using
   records

 - cvodes_bbd and ida_bbd take reasonably few arguments that NOT using
   records is justified

 - idas_bbd take so many arguments that records would be useful.