Add the Sparse Linear Solver (sls) data structures.
For Sundials-2.5.0, these data structures are simply not compiled. They are
only needed in Sundials-2.6.0 for the (optional) KLU and SuperLUMT linear
solvers.

Types are extracted from dls.ml to dls_impl.ml so that they can be used
in the convert functions of Sls.SparseMatrix.

Tidy up ocaml/linear examples and add a testing (diff) target.