New exception: NotImplementedBySundialsVersion
This exception is to be raised for features introduced in Sundials 2.6.0
which did not exist in earlier versions (and likewise for later versions).

The idea is that one can compile the current version of Sundials/ML (2.6.0)
against earlier versions of Sundials (e.g., 2.5.0). Newer features remain
visible in the interface (to minimise the complications of conditional
compilation in OCaml), but raise exceptions if invoked.