Big jactimesvec/precfn separation.
The Jacobian times vector function is conceptually distinct from the
preconditioner setup and solve functions. This commit separates them more
clearly in the session data structure.

This change makes it possible to use the BBD preconditioner with a Jacobian
times vector function. It also prepares the way for the upcoming addition of
the CVSpilsSetPreconditionerBS and CVSpilsSetJacTimesVecFnBS functions (as
well as their Ida* counterparts).