Revert "Keep building even if some examples fail."
I hear you, Tim.  You don't want to use -k.  I can have the Makefile do
it for you :)

Well, effectively.  The new code checks if a subdirectory's
tests.opt.log has been specified on the command line, and if so ignores
errors in those targets.  The errors are re-detected when the top-level
tests.opt.log's recipe runs.

The problem with the change in 91f6503c4b2a643b2f16335352fc359b7d7fb49a
is that it masks errors when you run the examples selectively.  For
instance, we'd like to be able to do

  cd examples
  make cvode/serial/tests.opt.log

and have the exit status reflect the outcome of the tests in that
subdirectory, because this can cut down the total running time of the
tests quite a bit.  Obliterating the grep at the end of the make target
would rule out this usage.