Read output of time(1) via a named file
perf used to read the output of time(1) through a pipe, which time(1)
wrote to its stderr.  But cvsHessian_ASA_FSA was having intermittent
failures, and when the failure was triggered, perf was crashing because
stderr contained unexpected output.

Named files are more annoying to handle and can introduce security
issues, but I don't think we need to be too paranoid about races in this
case.