<solver>.Direct/Iterative -> <solver>.Dls/Spils
In the solver modules (Cvode/Cvodes/Ida/Idas/Kinsol/Arkode), keep the
old submodule names for linear solvers. This is coherent with the
style of renaming CVSpilsSetJacTimes to Cvode.Spils.set_jac_times.

It is slightly incoherent with the module names in Lsolver (Direct and
Iterative), but this only reflects an incoherence in the underlying
library after the changes in 3.x.