Nasty GC bug with jacmats / simpler interface
The matrix argument of the session Dls.solver functions was passed to
Sundials which uses it internally for calculations and cloning. To stop
this matrix from being prematurely GCed, a reference was held in the
session object together with the jacobian/mass function callbacks.
But, only a matrix content object (Matrix.Dense/Band/Sparse.t) was held,
so the matrix wrapper object (Matrix.t) could be garbage collected and
thereby free the underlying and active SUNMatrix.

The fix is to hold a reference to the Matrix.t. Furthermore, we hold the
reference in the generic direct linear solver (to which the session
holds a reference to prevent premature garbage collection), and pass
this matrix "behind the scenes" when creating a session-specific solver.
This deviates from the Sundials interface, but it has two advantages:
1. Users need not pass the matrix argument twice.
2. We need not propagate the matrix kind type argument
   (which anyway needs only to be validated when creating a generic
    linear solver).