Ensure that Sundials.c_init_module is always called
Unless we actually use a value from, and not just a submodule of, the
Sundials module, it seems that its top-level definitions are not
executed. This can be verified by running

    objdump -d prog.opt | grep 'callq.*camlSundials__entry'

In this case, the Sundials.c_init_module function is not called and the
NotImplementedBySundialsVersion exception is not registered for access
via SUNDIALS_EXN. In this case, unimplemented functions segfault
rather than raise the exception.