Shift Matrix.t context holder into low-level object
We must keep a reference to the original OCaml context for as long as we
use its underlying C pointer. For Matrix.t, the reference was originally
held at the OCaml level, but this makes it tricky to implement a wrap
method for an arbitrary SUNMatrix. The reference has now been shifted
into the C object in the same manner as for the content backlink.