Fix read after free in OCaml 5.0 garbage collection
"Back references" are used from C code to recover the OCaml value
associated with a C pointer. They are values allocated within the C heap
and registered with the OCaml garbage collector by the
sunml_sundials_malloc_value function. They must be manually destroyed by
calling sunml_sundials_free_value, which unregisters the value and frees
the heap memory.

This mechanism is used to associate OCaml session records with C session
pointers. The back reference is (i) created when the OCaml session is
created; (ii) stored in the OCaml session record, and; (iii) destroyed
by the session finalizer.

During testing under OCaml 5.0, some examples failed with segmentation
errors. Execution with valgrind revealed a read-after-free access in the
GC (mark_slice_darken) to a session back reference. The (unconfirmed)
hypothesis is that, in OCaml 5.0, the session finalizer is called before
the session fields are traversed. In this case, the back reference is
first destroyed and then accessed. The solution is simply to replace the
back reference field in the session record with a dummy value (Val_unit)
as soon as the back reference is destroyed.