Add finalizer immediately to wrapped nvectors
The allocation of an nvector registers a global root. In the case of
wrapped nvectors, it's the finalizer that unregisters the root.
Associating the finalizer to the value immediately simplifies reasoning
about the property "if registered then eventually unregistered", by
eliminating the need to reason about failures after nvector allocation
but before a finalizer is associated.